+++ src/api/rick-and-morty.service.ts
@@ -1,7 +1,6 @@
 import { HttpClient } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { map } from 'rxjs';
 
 @Injectable({
   providedIn: 'root'
