+++ src/app/app.component.css
@@ -1,0 +1,3 @@
+app-hero {
+  padding: 10%;
+}
