+++ src/app/app.component.html
@@ -1,6 +1,8 @@
 <main class="main">
-  <app-navbar />
-  <app-character-list />
+    <app-navbar />
+    <app-hero />
+    <app-character-list/>
+    <app-footer/>
 </main>
 
 <router-outlet />
